@@ -1,35 +1,47 @@
 {
-    "name": "linio/cache",
-    "description": "Provides a multi-layered caching abstraction",
-    "keywords": ["linio", "cache", "redis", "memcache"],
-    "type": "library",
-    "license": "BSD-3-Clause",
-    "require": {
-        "php": ">=7.1",
-        "psr/log": "^1.0",
-        "linio/util": "^2.0",
-        "linio/database": "^2.0|^3.0",
-        "predis/predis": "^1.0",
-        "doctrine/inflector": "^1.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": ">=5.0",
-        "contorion/aerospike-stub": "1.0.*",
-        "friendsofphp/php-cs-fixer": "^2.4"
-    },
-    "suggest": {
-        "ext-phpiredis": "Allows faster serialization and deserialization of the Redis protocol",
-        "ext-aerospike": "Allows usage of Aerospike adapter",
-        "ext-redis": "Allows usage of phpredis adapter"
-    },
-    "autoload": {
-        "psr-4": {
-            "Linio\\Component\\Cache\\": "src"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Linio\\Component\\Cache\\": "tests"
-        }
+  "name": "linio/cache",
+  "description": "Provides a multi-layered caching abstraction",
+  "keywords": ["linio", "cache", "redis", "memcache"],
+  "type": "library",
+  "license": "BSD-3-Clause",
+  "require": {
+    "php": "^7.4",
+    "psr/log": "^1.0",
+    "linio/util": "^3.0",
+    "linio/database": "^4.0",
+    "predis/predis": "^1.0",
+    "doctrine/inflector": "^1.0"
+  },
+  "require-dev": {
+    "contorion/aerospike-stub": "1.0.*",
+    "friendsofphp/php-cs-fixer": "^2.16",
+    "phpstan/phpstan": "^0.12",
+    "phpunit/phpunit": "^8.5"
+  },
+  "suggest": {
+    "ext-phpiredis": "Allows faster serialization and deserialization of the Redis protocol",
+    "ext-aerospike": "Allows usage of Aerospike adapter",
+    "ext-redis": "Allows usage of phpredis adapter"
+  },
+  "autoload": {
+    "psr-4": {
+      "Linio\\Component\\Cache\\": "src"
     }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Linio\\Component\\Cache\\": "tests"
+    }
+  },
+  "scripts": {
+    "lint": "php-cs-fixer fix --verbose --show-progress=estimating",
+    "lint:check": "php-cs-fixer fix --dry-run --verbose --show-progress=estimating",
+    "phpunit": "phpunit",
+    "phpstan": "phpstan analyze",
+    "test": [
+      "@lint:check",
+      "@phpunit",
+      "@phpstan"
+    ]
+  }
 }
